Some refactoring where MS-SRVS and MS-SCMR are moved to their own packages and common DCERPC structs are relocated to a dedicated file. Some changes to NDR serialization functions to better handle both null- and non-null-terminated strings. Fixed an issue with a lack SMB credits when using Kerberos auth. And some small changes to make certain functions public and addition of error codes
