+++ smb/dcerpc/msscmr/scmr_test.go
@@ -19,13 +19,16 @@
 // LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 // SOFTWARE.
-package dcerpc
+package msscmr
 
 import (
 	"bytes"
 	"encoding/hex"
 
 	"testing"
+
+	"github.com/jfjallid/go-smb/msdtyp"
+	"github.com/jfjallid/go-smb/smb/dcerpc"
 )
 
 func TestOpenSCManagerReq(t *testing.T) {
@@ -383,7 +386,7 @@ func TestQueryServiceConfigRes(t *testing.T) {
 
 func TestChangeServiceConfigReq(t *testing.T) {
 	// Simple test to verify that the packet structure is valid
-	pkt, err := hex.DecodeString("000000009b8a1554f3da95418cbe3c9f6be68d0bffffffff0400000000000000010000002c000000000000002c00000043003a005c00570069006e0064006f00770073005c00530079007300740065006d00330032005c0073007600630068006f00730074002e0065007800650020002d006b002000720064007800670072006f0075007000000000000000000000000000000000000000030000001000000000000000100000002e005c00610064006d0069006e006900730074007200610074006f007200000004000000200000003e3af69cb90cdb3b983dc9c3d7042d72e1981e344d226c6789a3237e184262e52000000005000000150000000000000015000000520065007400610069006c002000640065006d006f0020007300650072007600690063006500320000000000")
+	pkt, err := hex.DecodeString("000000009b8a1554f3da95418cbe3c9f6be68d0bffffffff0400000000000000010000002c000000000000002c00000043003a005c00570069006e0064006f00770073005c00530079007300740065006d00330032005c0073007600630068006f00730074002e0065007800650020002d006b002000720064007800670072006f0075007000000000000000000000000000000000000000020000001000000000000000100000002e005c00610064006d0069006e006900730074007200610074006f007200000003000000200000003e3af69cb90cdb3b983dc9c3d7042d72e1981e344d226c6789a3237e184262e52000000004000000150000000000000015000000520065007400610069006c002000640065006d006f0020007300650072007600690063006500320000000000")
 	if err != nil {
 		t.Fatal(err)
 	}
@@ -412,8 +415,8 @@ func TestChangeServiceConfigReq(t *testing.T) {
 	}
 
 	password := "secretpass"
-	uncPassword := ToUnicode(password + "\x00")
-	encPassword, err := encryptSecret(sessionKey, uncPassword)
+	uncPassword := msdtyp.ToUnicode(password + "\x00")
+	encPassword, err := dcerpc.EncryptSecretDes(sessionKey, uncPassword)
 	if err != nil {
 		t.Fatal(err)
 	}
@@ -498,12 +501,12 @@ func TestCloseServiceHandleRes(t *testing.T) {
 
 func TestRCreateServiceReq(t *testing.T) {
 	// Simple test to verify that the packet structure is valid
-	pkt, err := hex.DecodeString("000000008380de25416dfe4ab011e6b3fbf620980a000000000000000a0000004d00690073006300530056004300310032000000010000000a000000000000000a0000004d00690073006300530056004300310032000000ff010f001000000003000000000000001d000000000000001d00000043003a005c00770069006e0064006f00770073005c00740065006d0070005c006f006e006500640072006900760065002e006500780065000000000000000000000000000000000000000000030000000c000000000000000c0000004c006f00630061006c00530079007300740065006d0000000000000000000000")
+	pkt, err := hex.DecodeString("000000008a8940a6acf44c468919b58072337e120a000000000000000a0000004d00690073006300530056004300310032000000010000000a000000000000000a0000004d00690073006300530056004300310032000000ff010f001000000003000000000000001d000000000000001d00000043003a005c00570069006e0064006f00770073005c00740065006d0070005c006f006e006500640072006900760065002e006500780065000000000000000000000000000000000000000000020000000c000000000000000c0000004c006f00630061006c00530079007300740065006d0000000000000000000000")
 	if err != nil {
 		t.Fatal(err)
 	}
 
-	handle, err := hex.DecodeString("000000008380de25416dfe4ab011e6b3fbf62098")
+	handle, err := hex.DecodeString("000000008a8940a6acf44c468919b58072337e12")
 	if err != nil {
 		t.Fatal(err)
 	}
@@ -516,7 +519,7 @@ func TestRCreateServiceReq(t *testing.T) {
 		ServiceType:      0x10,
 		StartType:        3,
 		ErrorControl:     0,
-		BinaryPathName:   `C:\windows\temp\onedrive.exe`,
+		BinaryPathName:   `C:\Windows\temp\onedrive.exe`,
 		LoadOrderGroup:   "",
 		TagId:            0,
 		Dependencies:     "",
