+++ smb/dcerpc/utils.go
@@ -23,229 +23,11 @@
 package dcerpc
 
 import (
-	"bytes"
 	"crypto/des"
 	"encoding/binary"
 	"fmt"
-	"io"
 	"math/bits"
-	"unicode/utf16"
 )
-
-func FromUnicodeString(buf []byte) (res string, err error) {
-	buflen := len(buf)
-	if (buflen % 2) != 0 {
-		err = fmt.Errorf("Invalid Unicode (UTF-16-LE) string")
-		return
-	}
-
-	s := make([]uint16, buflen/2)
-	err = binary.Read(bytes.NewReader(buf), le, &s)
-	if err != nil {
-		return
-	}
-	return string(utf16.Decode(s)), nil
-}
-
-func FromUnicode(buf []byte) ([]byte, error) {
-	s, err := FromUnicodeString(buf)
-	return []byte(s), err
-}
-
-func ToUnicode(input string) []byte {
-	codePoints := utf16.Encode([]rune(input))
-	b := bytes.Buffer{}
-	binary.Write(&b, le, &codePoints)
-	return b.Bytes()
-}
-
-// Return the values needed to encode a unicode string according to NDR (except for the Ptrs and MaxCount which has to be added manually)
-func newUnicodeStr(s string) (offset uint32, actualCount uint32, paddlen int, buffer []byte) {
-	data := s + "\x00"
-	buffer = ToUnicode(data)
-	actualCount = uint32(len(buffer) / 2)
-	offset = 0
-	paddlen = (len(buffer) % 4) //Got to be 4 byte aligned
-	if paddlen != 0 {
-		paddlen = 4 - paddlen
-	}
-	return
-}
-
-// Write a conformant and varying string to the output stream
-func writeConformantVaryingString(w io.Writer, s string) (n int, err error) {
-	offset, count, paddlen, buffer := newUnicodeStr(s)
-	err = binary.Write(w, le, count) // MaxCount
-	if err != nil {
-		return
-	}
-	n += 4
-	err = binary.Write(w, le, offset)
-	if err != nil {
-		return
-	}
-	n += 4
-	err = binary.Write(w, le, count)
-	if err != nil {
-		return
-	}
-	n += 4
-	_, err = w.Write(buffer)
-	if err != nil {
-		return
-	}
-	n += len(buffer)
-	padd := make([]byte, paddlen)
-	_, err = w.Write(padd)
-	if err != nil {
-		return
-	}
-	n += paddlen
-	return
-}
-
-// Write a ptr to a conformant and varying string to the output stream
-func WriteConformantVaryingStringPtr(w io.Writer, s string, refid uint32) (n int, err error) {
-	var n2 int
-
-	if s == "" {
-		// Empty strings are represented as a NULL Ptr
-		n, err = w.Write([]byte{0, 0, 0, 0})
-		if err != nil {
-			log.Errorln(err)
-		}
-		return
-	}
-	if refid != 0 {
-		err = binary.Write(w, le, refid)
-		if err != nil {
-			return
-		}
-		n = 4
-	}
-	n2, err = writeConformantVaryingString(w, s)
-	n += n2
-	return
-}
-
-func writeConformantArray(w io.Writer, buf []byte) (n int, err error) {
-	err = binary.Write(w, le, uint32(len(buf))) // MaxCount
-	if err != nil {
-		log.Errorln(err)
-		return
-	}
-	n += 4
-	_, err = w.Write(buf)
-	if err != nil {
-		log.Errorln(err)
-		return
-	}
-	n += len(buf)
-	paddlen := (len(buf) % 4) //Got to be 4 byte aligned?
-	if paddlen != 0 {
-		paddlen = 4 - paddlen
-	}
-
-	padd := make([]byte, paddlen)
-	_, err = w.Write(padd)
-	if err != nil {
-		log.Errorln(err)
-		return
-	}
-	n += paddlen
-	return
-}
-
-func writeConformantArrayPtr(w io.Writer, buf []byte, refid uint32) (n int, err error) {
-	var n2 int
-
-	if len(buf) == 0 {
-		// Empty buffers are represented with a NULL Ptr
-		n, err = w.Write([]byte{0, 0, 0, 0})
-		if err != nil {
-			log.Errorln(err)
-		}
-		return
-	}
-	if refid != 0 {
-		err = binary.Write(w, le, refid)
-		if err != nil {
-			log.Errorln(err)
-			return
-		}
-		n = 4
-	}
-	n2, err = writeConformantArray(w, buf)
-	n += n2
-	return
-}
-
-func ReadConformantVaryingString(r *bytes.Reader) (s string, err error) {
-	// Read the Max count
-	var maxCount uint32
-	err = binary.Read(r, le, &maxCount)
-	if err != nil {
-		log.Errorln(err)
-		return
-	}
-	if maxCount == 0 {
-		// If maxCount is zero, we've likely encountered a null ptr
-		return
-	}
-	// Read the offset
-	var offset uint32
-	err = binary.Read(r, le, &offset)
-	if err != nil {
-		log.Errorln(err)
-		return
-	}
-	// Read the Actual count
-	var actualCount uint32
-	err = binary.Read(r, le, &actualCount)
-	if err != nil {
-		log.Errorln(err)
-		return
-	}
-	if offset > 0 {
-		_, err = r.Seek(int64(offset), io.SeekCurrent)
-		if err != nil {
-			log.Errorln(err)
-			return
-		}
-	}
-
-	if actualCount > 0 {
-		// Read the unicode string
-		unc := make([]byte, actualCount*2)
-		err = binary.Read(r, le, unc)
-		if err != nil {
-			log.Errorln(err)
-			return
-		}
-		// Check for terminating null byte
-		lastIndex := len(unc) - 2
-		if bytes.Compare(unc[lastIndex:], []byte{0, 0}) == 0 {
-			unc = unc[:lastIndex]
-		}
-		s, err = FromUnicodeString(unc)
-		if err != nil {
-			log.Errorln(err)
-			return
-		}
-		//s = s[:len(s)-1] // Skip terminating null character (already handled above)
-	}
-
-	paddLen := 4 - ((offset + actualCount*2) % 4)
-
-	if paddLen != 4 {
-		_, err = r.Seek(int64(paddLen), io.SeekCurrent)
-		if err != nil {
-			log.Errorln(err)
-			return
-		}
-	}
-	return
-}
 
 // A couple of helper functions defined in MS-LSAD 5.1.2 for Secret Encryption
 
@@ -280,7 +62,7 @@ func transformKey(input []byte) []byte {
 }
 
 // MS-LSAD 5.1.2 and 5.1.3 combined
-func encryptSecret(key, input []byte) (ciphertext []byte, err error) {
+func EncryptSecretDes(key, input []byte) (ciphertext []byte, err error) {
 	if len(key) != 16 {
 		err = fmt.Errorf("Invalid key size for MS-LSAD secret encryption. Expected a 16 byte key!")
 		log.Errorln(err)
